refactor email settings
