+++ aitutor/email/__init__.py
@@ -1,0 +1,11 @@
+from .azure import AzureEmailSender
+from .base import EmailSender
+from .gmail import GmailEmailSender
+from .smtp import SmtpEmailSender
+
+__all__ = [
+    "EmailSender",
+    "SmtpEmailSender",
+    "AzureEmailSender",
+    "GmailEmailSender",
+]
